remove execution-graph-utils & execution-status-graph file. remove visual and step function graph rendering from execution status. move getEventDetails function from execution-graph-utils to execution.js
